refactor repository structure and update imports for clarity
adapters are placed as infrastructure
